# [perf] feat: Padding before batch post-process in agent-loop to save time